Add files via upload
- Correction des noms qui s'affichait plus
- La moyenne s'affiche désormais correctement

+++ POO_JeuNombreADeviner/include/Joueur.h
@@ -58,7 +58,7 @@ class CJoueur {
     string Nom();
 
     void ResultatsMoy(float &moy) {
-        moy = nbTentatives / nbPartiesJouees;
+        moy = (float) nbTentatives / nbPartiesJouees;
     }
 
     private :
